refactor: modifica caminho de um import
