+++ src/services/patients/getAllPatients.svc.ts
@@ -1,5 +1,5 @@
 import { IPatient } from "../../interfaces/patient.interface";
-import { IPatientRepository } from "../../repositories/implementations/patients";
+import { IPatientRepository } from "../../repositories/implementations/IPatientRepository";
 
 class GetAllPatientsSvc {
   constructor(private patientRepository: IPatientRepository) {}
